requirements.txt
telethon
fastapi
uvicorn
requests
pydantic
gspread
google-auth
